+++ gen-src/day8/re_frame_10x/inlined_deps/reagent/v1v0v0/reagent/impl/protocols.cljs
@@ -1,7 +1,0 @@
-(ns ^{:mranderson/inlined true} day8.re-frame-10x.inlined-deps.reagent.v1v0v0.reagent.impl.protocols)
-
-(defprotocol Compiler
-  (get-id [this])
-  (as-element [this x])
-  (make-element [this argv component jsprops first-child]))
-