+++ gen-src/day8/re_frame_10x/inlined_deps/reagent/v1v0v0/reagent/interop.cljs
@@ -1,2 +1,0 @@
-(ns ^{:mranderson/inlined true} day8.re-frame-10x.inlined-deps.reagent.v1v0v0.reagent.interop
-  (:require-macros [day8.re-frame-10x.inlined-deps.reagent.v1v0v0.reagent.interop]))