+++ gen-src/day8/re_frame_10x/inlined_deps/reagent/v1v2v0/reagent/interop.clj
@@ -1,0 +1,20 @@
+(ns ^{:mranderson/inlined true} day8.re-frame-10x.inlined-deps.reagent.v1v2v0.reagent.interop)
+
+; taken from cljs.core
+; https://github.com/binaryage/cljs-oops/issues/14
+(defmacro unchecked-aget
+  ([array idx]
+   (list 'js* "(~{}[~{}])" array idx))
+  ([array idx & idxs]
+   (let [astr (apply str (repeat (count idxs) "[~{}]"))]
+     `(~'js* ~(str "(~{}[~{}]" astr ")") ~array ~idx ~@idxs))))
+
+; taken from cljs.core
+; https://github.com/binaryage/cljs-oops/issues/14
+(defmacro unchecked-aset
+  ([array idx val]
+   (list 'js* "(~{}[~{}] = ~{})" array idx val))
+  ([array idx idx2 & idxv]
+   (let [n (dec (count idxv))
+         astr (apply str (repeat n "[~{}]"))]
+     `(~'js* ~(str "(~{}[~{}][~{}]" astr " = ~{})") ~array ~idx ~idx2 ~@idxv))))