+++ src/day8/re_frame_10x/fx/window.cljs
@@ -3,7 +3,7 @@
    [goog.object                                                  :as gobj]
    [goog.string                                                  :as gstring]
    [clojure.string                                               :as string]
-   [day8.re-frame-10x.inlined-deps.re-frame.v1v1v2.re-frame.core :as rf]))
+   [day8.re-frame-10x.inlined-deps.re-frame.v1v3v0.re-frame.core :as rf]))
 
 (defn m->str
   [m]