+++ src/day8/re_frame_10x/panels/fx/views.cljs
@@ -3,7 +3,7 @@
    [day8.re-frame-10x.components.re-com                          :as rc]
    [day8.re-frame-10x.inlined-deps.garden.v1v3v10.garden.units   :refer [percent]]
    [day8.re-frame-10x.inlined-deps.spade.git-sha-93ef290.core       :refer [defclass]]
-   [day8.re-frame-10x.inlined-deps.re-frame.v1v1v2.re-frame.core :as rf]
+   [day8.re-frame-10x.inlined-deps.re-frame.v1v3v0.re-frame.core :as rf]
    [day8.re-frame-10x.components.buttons                         :as buttons]
    [day8.re-frame-10x.components.cljs-devtools                   :as cljs-devtools]
    [day8.re-frame-10x.material                                   :as material]