+++ src/day8/re_frame_10x/panels/settings/views.cljs
@@ -1,6 +1,6 @@
 (ns day8.re-frame-10x.panels.settings.views
   (:require
-   [day8.re-frame-10x.inlined-deps.re-frame.v1v1v2.re-frame.core  :as rf]
+   [day8.re-frame-10x.inlined-deps.re-frame.v1v3v0.re-frame.core  :as rf]
    [day8.re-frame-10x.inlined-deps.garden.v1v3v10.garden.units    :refer [px* px-]]
    [day8.re-frame-10x.inlined-deps.garden.v1v3v10.garden.compiler :refer [render-css]]
    [day8.re-frame-10x.components.buttons                          :as buttons]