+++ src/day8/re_frame_10x/panels/subs/views.cljs
@@ -3,7 +3,7 @@
    [day8.re-frame-10x.components.re-com                          :refer [handler-fn]])
   (:require
    [clojure.data]
-   [day8.re-frame-10x.inlined-deps.re-frame.v1v1v2.re-frame.core :as rf]
+   [day8.re-frame-10x.inlined-deps.re-frame.v1v3v0.re-frame.core :as rf]
    [day8.re-frame-10x.inlined-deps.spade.git-sha-93ef290.core    :refer [defclass]]
    [day8.re-frame-10x.inlined-deps.garden.v1v3v10.garden.units   :refer [px percent]]
    [day8.re-frame-10x.inlined-deps.garden.v1v3v10.garden.color   :as color]