+++ src/day8/re_frame_10x/panels/traces/events.cljs
@@ -2,7 +2,7 @@
   (:require
    [clojure.set                                                  :as set]
    [clojure.string                                               :as string]
-   [day8.re-frame-10x.inlined-deps.re-frame.v1v1v2.re-frame.core :as rf]
+   [day8.re-frame-10x.inlined-deps.re-frame.v1v3v0.re-frame.core :as rf]
    [day8.re-frame-10x.fx.local-storage                           :as local-storage]))
 
 (rf/reg-event-db