+++ src/day8/re_frame_10x/panels/traces/subs.cljs
@@ -1,7 +1,7 @@
 (ns day8.re-frame-10x.panels.traces.subs
   (:require
    [clojure.string                                               :as string]
-   [day8.re-frame-10x.inlined-deps.re-frame.v1v1v2.re-frame.core :as rf]
+   [day8.re-frame-10x.inlined-deps.re-frame.v1v3v0.re-frame.core :as rf]
    [day8.re-frame-10x.tools.metamorphic                          :as metam]
    [day8.re-frame-10x.navigation.epochs.subs                     :as epochs.subs]
    [day8.re-frame-10x.panels.settings.subs                       :as settings.subs]